[#1278] Create CalculatorStorage
