+++ client/src/commonMain/kotlin/com/oztechan/ccc/client/di/module/submodule/StorageModule.kt
@@ -1,11 +1,14 @@
 package com.oztechan.ccc.client.di.module.submodule
 
-import com.oztechan.ccc.client.storage.AppStorage
-import com.oztechan.ccc.client.storage.AppStorageImpl
+import com.oztechan.ccc.client.storage.app.AppStorage
+import com.oztechan.ccc.client.storage.app.AppStorageImpl
+import com.oztechan.ccc.client.storage.calculator.CalculatorStorage
+import com.oztechan.ccc.client.storage.calculator.CalculatorStorageImpl
 import org.koin.core.module.dsl.bind
 import org.koin.core.module.dsl.singleOf
 import org.koin.dsl.module
 
 internal val storageModule = module {
     singleOf(::AppStorageImpl) { bind<AppStorage>() }
+    singleOf(::CalculatorStorageImpl) { bind<CalculatorStorage>() }
 }
