+++ client/src/commonMain/kotlin/com/oztechan/ccc/client/di/module/submodule/ViewModelModule.kt
@@ -15,9 +15,9 @@ import org.koin.core.qualifier.Qualifier
 import org.koin.dsl.module
 
 internal val viewModelModule = module {
-    viewModelDefinition { SettingsViewModel(get(), get(), get(), get(), get(), get(), get(), get()) }
+    viewModelDefinition { SettingsViewModel(get(), get(), get(), get(), get(), get(), get(), get(), get()) }
     viewModelDefinition { MainViewModel(get(), get(), get(), get(), get(), get()) }
-    viewModelDefinition { CurrenciesViewModel(get(), get(), get(), get()) }
+    viewModelDefinition { CurrenciesViewModel(get(), get(), get(), get(), get()) }
     viewModelDefinition { CalculatorViewModel(get(), get(), get(), get(), get(), get()) }
     viewModelDefinition { SelectCurrencyViewModel(get()) }
     viewModelDefinition { AdRemoveViewModel(get()) }
