+++ client/src/commonMain/kotlin/com/oztechan/ccc/client/repository/ad/AdRepositoryImpl.kt
@@ -1,7 +1,7 @@
 package com.oztechan.ccc.client.repository.ad
 
 import com.oztechan.ccc.client.model.Device
-import com.oztechan.ccc.client.storage.AppStorage
+import com.oztechan.ccc.client.storage.app.AppStorage
 import com.oztechan.ccc.client.util.isRewardExpired
 import com.oztechan.ccc.config.service.ad.AdConfigService
