+++ client/src/commonMain/kotlin/com/oztechan/ccc/client/storage/app/AppStorage.kt
@@ -1,15 +1,11 @@
-package com.oztechan.ccc.client.storage
+package com.oztechan.ccc.client.storage.app
 
 interface AppStorage {
     var firstRun: Boolean
-
-    var currentBase: String
 
     var appTheme: Int
 
     var adFreeEndDate: Long
 
     var sessionCount: Long
-
-    var precision: Int
 }
