+++ client/src/commonMain/kotlin/com/oztechan/ccc/client/storage/calculator/CalculatorStorage.kt
@@ -1,0 +1,7 @@
+package com.oztechan.ccc.client.storage.calculator
+
+interface CalculatorStorage {
+    var currentBase: String
+
+    var precision: Int
+}
