+++ client/src/commonMain/kotlin/com/oztechan/ccc/client/viewmodel/calculator/CalculatorViewModel.kt
@@ -18,7 +18,7 @@ import com.oztechan.ccc.client.mapper.toUIModelList
 import com.oztechan.ccc.client.model.Currency
 import com.oztechan.ccc.client.model.RateState
 import com.oztechan.ccc.client.repository.ad.AdRepository
-import com.oztechan.ccc.client.storage.AppStorage
+import com.oztechan.ccc.client.storage.calculator.CalculatorStorage
 import com.oztechan.ccc.client.util.MAXIMUM_FLOATING_POINT
 import com.oztechan.ccc.client.util.calculateResult
 import com.oztechan.ccc.client.util.getCurrencyConversionByRate
@@ -50,7 +50,7 @@ import kotlinx.coroutines.launch
 
 @Suppress("TooManyFunctions")
 class CalculatorViewModel(
-    private val appStorage: AppStorage,
+    private val calculatorStorage: CalculatorStorage,
     private val backendApiService: BackendApiService,
     private val currencyDataSource: CurrencyDataSource,
     private val offlineRatesDataSource: OfflineRatesDataSource,
@@ -71,7 +71,7 @@ class CalculatorViewModel(
 
     init {
         _state.update {
-            copy(base = appStorage.currentBase, input = "")
+            copy(base = calculatorStorage.currentBase, input = "")
         }
 
         state.map { it.base }
@@ -106,7 +106,7 @@ class CalculatorViewModel(
     private fun getRates() = data.rates?.let {
         calculateConversions(it, RateState.Cached(it.date))
     } ?: viewModelScope.launch {
-        runCatching { backendApiService.getRates(appStorage.currentBase) }
+        runCatching { backendApiService.getRates(calculatorStorage.currentBase) }
             .onFailure(::getRatesFailed)
             .onSuccess(::getRatesSuccess)
     }
@@ -124,7 +124,7 @@ class CalculatorViewModel(
     private fun getRatesFailed(t: Throwable) = viewModelScope.launchIgnored {
         Logger.w(t) { "CalculatorViewModel getRatesFailed" }
         offlineRatesDataSource.getOfflineRatesByBase(
-            appStorage.currentBase
+            calculatorStorage.currentBase
         )?.let {
             calculateConversions(it, RateState.Offline(it.date))
         } ?: run {
@@ -147,7 +147,7 @@ class CalculatorViewModel(
     private fun calculateOutput(input: String) = viewModelScope.launch {
         data.parser
             .calculate(input.toSupportedCharacters(), MAXIMUM_FLOATING_POINT)
-            .mapTo { if (isFinite()) getFormatted(appStorage.precision) else "" }
+            .mapTo { if (isFinite()) getFormatted(calculatorStorage.precision) else "" }
             .whether(
                 { output -> output.length <= MAXIMUM_OUTPUT },
                 { input.length <= MAXIMUM_INPUT }
@@ -173,7 +173,7 @@ class CalculatorViewModel(
         copy(
             currencyList = _state.value.currencyList.onEach {
                 it.rate = rates.calculateResult(it.name, _state.value.output)
-                    .getFormatted(appStorage.precision)
+                    .getFormatted(calculatorStorage.precision)
                     .toStandardDigits()
             },
             rateState = rateState,
@@ -183,7 +183,7 @@ class CalculatorViewModel(
 
     private fun currentBaseChanged(newBase: String, shouldTrack: Boolean = false) = viewModelScope.launchIgnored {
         data.rates = null
-        appStorage.currentBase = newBase
+        calculatorStorage.currentBase = newBase
         _state.update {
             copy(
                 loading = true,
@@ -237,7 +237,7 @@ class CalculatorViewModel(
             _effect.emit(
                 CalculatorEffect.ShowRate(
                     currency.getCurrencyConversionByRate(
-                        appStorage.currentBase,
+                        calculatorStorage.currentBase,
                         data.rates
                     ),
                     currency.name
