+++ client/src/commonTest/kotlin/com/oztechan/ccc/client/viewmodel/CurrenciesViewModelTest.kt
@@ -7,7 +7,8 @@ import com.oztechan.ccc.analytics.AnalyticsManager
 import com.oztechan.ccc.analytics.model.UserProperty
 import com.oztechan.ccc.client.mapper.toUIModel
 import com.oztechan.ccc.client.repository.ad.AdRepository
-import com.oztechan.ccc.client.storage.AppStorage
+import com.oztechan.ccc.client.storage.app.AppStorage
+import com.oztechan.ccc.client.storage.calculator.CalculatorStorage
 import com.oztechan.ccc.client.viewmodel.currencies.CurrenciesEffect
 import com.oztechan.ccc.client.viewmodel.currencies.CurrenciesViewModel
 import com.oztechan.ccc.common.datasource.currency.CurrencyDataSource
@@ -40,11 +41,14 @@ import com.oztechan.ccc.common.model.Currency as CommonCurrency
 internal class CurrenciesViewModelTest : BaseViewModelTest<CurrenciesViewModel>() {
 
     override val subject: CurrenciesViewModel by lazy {
-        CurrenciesViewModel(appStorage, currencyDataSource, adRepository, analyticsManager)
+        CurrenciesViewModel(appStorage, calculatorStorage, currencyDataSource, adRepository, analyticsManager)
     }
 
     @Mock
     private val appStorage = mock(classOf<AppStorage>())
+
+    @Mock
+    private val calculatorStorage = mock(classOf<CalculatorStorage>())
 
     @Mock
     private val currencyDataSource = mock(classOf<CurrencyDataSource>())
@@ -79,7 +83,7 @@ internal class CurrenciesViewModelTest : BaseViewModelTest<CurrenciesViewModel>(
             .invocation { firstRun }
             .thenReturn(false)
 
-        given(appStorage)
+        given(calculatorStorage)
             .invocation { currentBase }
             .thenReturn(clientCurrency.name)
     }
@@ -155,7 +159,7 @@ internal class CurrenciesViewModelTest : BaseViewModelTest<CurrenciesViewModel>(
 
     @Test
     fun `don't show FewCurrency effect if there is MINIMUM_ACTIVE_CURRENCY and not firstRun`() {
-        given(appStorage)
+        given(calculatorStorage)
             .invocation { currentBase }
             .thenReturn("") // in order to get ChangeBase effect, have to have an effect to finish test
 
@@ -181,7 +185,7 @@ internal class CurrenciesViewModelTest : BaseViewModelTest<CurrenciesViewModel>(
             .invocation { firstRun }
             .thenReturn(true)
 
-        given(appStorage)
+        given(calculatorStorage)
             .invocation { currentBase }
             .thenReturn("") // in order to get ChangeBase effect, have to have an effect to finish test
 
@@ -262,7 +266,7 @@ internal class CurrenciesViewModelTest : BaseViewModelTest<CurrenciesViewModel>(
                 }
             )
 
-        given(appStorage)
+        given(calculatorStorage)
             .invocation { currentBase }
             .thenReturn("")
 
@@ -271,7 +275,7 @@ internal class CurrenciesViewModelTest : BaseViewModelTest<CurrenciesViewModel>(
             assertEquals(firstActiveBase, it.newBase)
         }
 
-        verify(appStorage)
+        verify(calculatorStorage)
             .invocation { currentBase = firstActiveBase }
             .wasInvoked()
     }
@@ -289,7 +293,7 @@ internal class CurrenciesViewModelTest : BaseViewModelTest<CurrenciesViewModel>(
                 }
             )
 
-        given(appStorage)
+        given(calculatorStorage)
             .invocation { currentBase }
             .thenReturn(commonCurrency.name) // not active one
 
@@ -298,7 +302,7 @@ internal class CurrenciesViewModelTest : BaseViewModelTest<CurrenciesViewModel>(
             assertEquals(commonCurrency2.name, it.newBase)
         }
 
-        verify(appStorage)
+        verify(calculatorStorage)
             .invocation { currentBase = commonCurrency2.name }
             .wasInvoked()
     }
@@ -310,7 +314,7 @@ internal class CurrenciesViewModelTest : BaseViewModelTest<CurrenciesViewModel>(
             .invocation { firstRun }
             .thenReturn(false)
 
-        given(appStorage)
+        given(calculatorStorage)
             .invocation { currentBase }
             .thenReturn("EUR")
